BOSA-81: migrate Decidim::Initiatives::CreateInitiative to extension
We want to modify only few methods, so we don't need to include the
class inside this extension. For the rich editor context, we don't want
to strip_tags for initiative's description.
